To_Lexa/backend/config.py
import os
import logging
from supabase import create_client, Client
from openai import OpenAI
from typing import Optional
from dotenv import load_dotenv

# --- Конфигурация логирования ---
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

load_dotenv()
# --- Конфигурация и переменные окружения ---
SUPABASE_URL = os.getenv("SUPABASE_URL", "").strip()
SUPABASE_KEY = os.getenv("SUPABASE_KEY", "").strip()
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "").strip()
SECRET_KEY = os.getenv("SECRET_KEY", "")


# --- Инициализация клиентов ---
supabase: Optional[Client] = None
openai_client: Optional[OpenAI] = None

if SUPABASE_URL and SUPABASE_KEY:
    supabase = create_client(SUPABASE_URL, SUPABASE_KEY)

if OPENAI_API_KEY:
    openai_client = OpenAI(api_key=OPENAI_API_KEY)