To_Lexa/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.3
openai==1.3.5
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyJWT